feat: merge `barretenberg_static_lib` and `barretenberg_wasm`
