+++ acvm_backend_barretenberg/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "barretenberg_wasm"
+name = "acvm-backend-barretenberg"
 version = "0.1.0"
 authors.workspace = true
 license.workspace = true
@@ -9,10 +9,16 @@ rust-version.workspace = true
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-wasmer = { version = "*", default-features = false }
+cfg-if = "1.0.0"
 common = { path = "../common", default-features = false }
-rust-embed = { version = "6.6.0", features = ["debug-embed", "interpolate-folder-path", "include-exclude"] }
-getrandom = "0.2"
+
+# Native
+barretenberg-sys = { version = "0.1.1", optional = true }
+
+# Wasm
+wasmer = { version = "*", optional = true, default-features = false }
+rust-embed = { version = "6.6.0", optional = true, features = ["debug-embed", "interpolate-folder-path", "include-exclude"] }
+getrandom = { version = "0.2", optional = true }
 
 [build-dependencies]
 pkg-config = "0.3"
@@ -22,8 +28,16 @@ blake2 = "0.9.1"
 tempfile = "*"
 
 [features]
-default = [
-    # "acvm/bn254",
+default = ["native"]
+
+native=[
+    "dep:barretenberg-sys",
+    "common/std",
+]
+wasm = [
+    "wasmer",
+    "dep:rust-embed",
+    "dep:getrandom",
     "common/std",
     "wasmer/sys-default",
     "wasmer/cranelift",
@@ -32,7 +46,8 @@ default = [
     "wasmer/default-universal",
 ]
 js = [
-
-    # "acvm/bn254", 
+    "wasmer",
+    "dep:rust-embed",
+    "dep:getrandom",
     "wasmer/js-default",
 ]
