+++ acvm_backend_barretenberg/src/acvm_interop/mod.rs
@@ -1,7 +1,7 @@
-pub mod proof_system;
-pub mod pwg;
-
+mod proof_system;
+mod pwg;
 mod smart_contract;
+
 pub struct Plonk;
 
 impl common::acvm::Backend for Plonk {}
