+++ acvm_backend_barretenberg/src/acvm_interop/pwg/black_box_functions.rs
