+++ acvm_backend_barretenberg/src/acvm_interop/pwg/merkle.rs
