+++ acvm_backend_barretenberg/src/blake2s.rs
@@ -4,10 +4,30 @@ use super::Barretenberg;
 
 impl Barretenberg {
     /// Hashes to a bn254 scalar field element using blake2s
-    pub fn hash_to_field(&mut self, input: &[u8]) -> FieldElement {
-        let result_prt = barretenberg_sys::blake2s::hash_to_field(input);
+    pub(crate) fn hash_to_field(&mut self, input: &[u8]) -> FieldElement {
+        cfg_if::cfg_if! {
+            if #[cfg(feature = "native")] {
+                let result_bytes = barretenberg_sys::blake2s::hash_to_field(input);
+            } else {
+                use wasmer::Value;
 
-        FieldElement::from_be_bytes_reduce(&result_prt)
+                let input_ptr = self.allocate(input); // 0..32
+
+                let result_ptr = Value::I32(0);
+
+                // Not sure why this is needed to send to WASM
+                // It seems to be sent twice?
+                let data_len = Value::I32(input.len() as i32);
+
+                self.call_multiple("blake2s_to_field", vec![&input_ptr, &data_len, &result_ptr]);
+
+                self.free(input_ptr);
+
+                // result_ptr is 0 and the output is 32 bytes long
+                let result_bytes = self.slice_memory(0, 32);
+            }
+        }
+        FieldElement::from_be_bytes_reduce(&result_bytes)
     }
 }
