+++ acvm_backend_barretenberg/src/composer.rs
@@ -1,33 +1,164 @@
 use super::pippenger::Pippenger;
-use super::Barretenberg;
 use common::barretenberg_structures::*;
 use common::crs::CRS;
 use common::proof;
-use wasmer::Value;
 
-pub struct StandardComposer {
-    barretenberg: Barretenberg,
-    pippenger: Pippenger,
-    crs: CRS,
-    constraint_system: ConstraintSystem,
-}
+cfg_if::cfg_if! {
+    if #[cfg(feature = "native")] {
 
-impl StandardComposer {
-    pub fn new(constraint_system: ConstraintSystem) -> StandardComposer {
-        let mut barretenberg = Barretenberg::new();
+        pub(crate) struct StandardComposer {
+            pippenger: Pippenger,
+            crs: CRS,
+            constraint_system: ConstraintSystem,
+        }
 
-        let circuit_size =
-            StandardComposer::get_circuit_size(&mut barretenberg, &constraint_system);
+        fn pow2ceil(v: u32) -> u32 {
+            v.next_power_of_two()
+        }
 
-        let crs = CRS::new(circuit_size as usize);
+        impl StandardComposer {
+            pub(crate) fn new(constraint_system: ConstraintSystem) -> StandardComposer {
+                let circuit_size = StandardComposer::get_circuit_size(&constraint_system);
 
-        let pippenger = Pippenger::new(&crs.g1_data, &mut barretenberg);
+                let crs = CRS::new(circuit_size as usize + 1);
 
-        StandardComposer {
-            barretenberg,
-            pippenger,
-            crs,
-            constraint_system,
+                let pippenger = Pippenger::new(&crs.g1_data);
+
+                StandardComposer {
+                    pippenger,
+                    crs,
+                    constraint_system,
+                }
+            }
+
+            // XXX: There seems to be a bug in the C++ code
+            // where it causes a `HeapAccessOutOfBound` error
+            // for certain circuit sizes.
+            //
+            // This method calls the WASM for the circuit size
+            // if an error is returned, then the circuit size is defaulted to 2^19.
+            //
+            // This method is primarily used to determine how many group
+            // elements we need from the CRS. So using 2^19 on an error
+            // should be an overestimation.
+            pub(crate) fn get_circuit_size(constraint_system: &ConstraintSystem) -> u32 {
+                let num_gates;
+                unsafe {
+                    num_gates = barretenberg_sys::composer::get_total_circuit_size(
+                        constraint_system.to_bytes().as_slice().as_ptr(),
+                    );
+                }
+                pow2ceil(num_gates + StandardComposer::NUM_RESERVED_GATES)
+            }
+
+            pub(crate) fn get_exact_circuit_size(constraint_system: &ConstraintSystem) -> u32 {
+                unsafe {
+                    barretenberg_sys::composer::get_exact_circuit_size(
+                        constraint_system.to_bytes().as_slice().as_ptr(),
+                    )
+                }
+            }
+        }
+    } else {
+        use super::Barretenberg;
+
+        pub(crate) struct StandardComposer {
+            barretenberg: Barretenberg,
+            pippenger: Pippenger,
+            crs: CRS,
+            constraint_system: ConstraintSystem,
+        }
+
+        fn pow2ceil(v: u32) -> u32 {
+            let mut p = 1;
+            while p < v {
+                p <<= 1;
+            }
+            p
+        }
+
+        impl StandardComposer {
+            pub(crate) fn new(constraint_system: ConstraintSystem) -> StandardComposer {
+                let mut barretenberg = Barretenberg::new();
+
+                let circuit_size =
+                    StandardComposer::get_circuit_size(&mut barretenberg, &constraint_system);
+
+                let crs = CRS::new(circuit_size as usize);
+
+                let pippenger = Pippenger::new(&crs.g1_data, &mut barretenberg);
+
+                StandardComposer {
+                    barretenberg,
+                    pippenger,
+                    crs,
+                    constraint_system,
+                }
+            }
+
+            // XXX: There seems to be a bug in the C++ code
+            // where it causes a `HeapAccessOutOfBound` error
+            // for certain circuit sizes.
+            //
+            // This method calls the WASM for the circuit size
+            // if an error is returned, then the circuit size is defaulted to 2^19.
+            //
+            // This method is primarily used to determine how many group
+            // elements we need from the CRS. So using 2^19 on an error
+            // should be an overestimation.
+            pub(crate) fn get_circuit_size(
+                barretenberg: &mut Barretenberg,
+                constraint_system: &ConstraintSystem,
+            ) -> u32 {
+                let cs_buf = constraint_system.to_bytes();
+                let cs_ptr = barretenberg.allocate(&cs_buf);
+
+                let func = barretenberg
+                    .instance
+                    .exports
+                    .get_function("acir_proofs_get_total_circuit_size")
+                    .unwrap();
+
+                let params: Vec<_> = vec![cs_ptr.clone()];
+                match func.call(&params) {
+                    Ok(vals) => {
+                        let i32_bytes = vals.first().cloned().unwrap().unwrap_i32().to_be_bytes();
+                        let u32_val = u32::from_be_bytes(i32_bytes);
+                        barretenberg.free(cs_ptr);
+                        pow2ceil(u32_val + StandardComposer::NUM_RESERVED_GATES)
+                    }
+                    Err(_) => {
+                        unreachable!("failed on acir_proofs_get_total_circuit_size call");
+                    }
+                }
+            }
+
+            pub(crate) fn get_exact_circuit_size(
+                barretenberg: &mut Barretenberg,
+                constraint_system: &ConstraintSystem,
+            ) -> u32 {
+                let cs_buf = constraint_system.to_bytes();
+                let cs_ptr = barretenberg.allocate(&cs_buf);
+
+                let func = barretenberg
+                    .instance
+                    .exports
+                    .get_function("acir_proofs_get_exact_circuit_size")
+                    .unwrap();
+
+                let params: Vec<_> = vec![cs_ptr.clone()];
+                match func.call(&params) {
+                    Ok(vals) => {
+                        let i32_bytes = vals.first().cloned().unwrap().unwrap_i32().to_be_bytes();
+                        let u32_val = u32::from_be_bytes(i32_bytes);
+                        barretenberg.free(cs_ptr);
+                        u32_val
+                    }
+                    Err(_) => {
+                        unreachable!("failed on acir_proofs_get_exact_circuit_size call");
+                    }
+                }
+            }
         }
     }
 }
@@ -35,154 +166,176 @@ impl StandardComposer {
 impl StandardComposer {
     const NUM_RESERVED_GATES: u32 = 4; // this must be >= num_roots_cut_out_of_vanishing_polynomial (found under prover settings in barretenberg)
 
-    // XXX: There seems to be a bug in the C++ code
-    // where it causes a `HeapAccessOutOfBound` error
-    // for certain circuit sizes.
-    //
-    // This method calls the WASM for the circuit size
-    // if an error is returned, then the circuit size is defaulted to 2^19.
-    //
-    // This method is primarily used to determine how many group
-    // elements we need from the CRS. So using 2^19 on an error
-    // should be an overestimation.
-    pub fn get_circuit_size(
-        barretenberg: &mut Barretenberg,
-        constraint_system: &ConstraintSystem,
-    ) -> u32 {
-        let cs_buf = constraint_system.to_bytes();
-        let cs_ptr = barretenberg.allocate(&cs_buf);
-
-        let func = barretenberg
-            .instance
-            .exports
-            .get_function("acir_proofs_get_total_circuit_size")
-            .unwrap();
-
-        let params: Vec<_> = vec![cs_ptr.clone()];
-        match func.call(&params) {
-            Ok(vals) => {
-                let i32_bytes = vals.first().cloned().unwrap().unwrap_i32().to_be_bytes();
-                let u32_val = u32::from_be_bytes(i32_bytes);
-                barretenberg.free(cs_ptr);
-                pow2ceil(u32_val + StandardComposer::NUM_RESERVED_GATES)
-            }
-            Err(_) => {
-                unreachable!("failed on acir_proofs_get_total_circuit_size call");
-            }
-        }
-    }
-
-    pub fn get_exact_circuit_size(
-        barretenberg: &mut Barretenberg,
-        constraint_system: &ConstraintSystem,
-    ) -> u32 {
-        let cs_buf = constraint_system.to_bytes();
-        let cs_ptr = barretenberg.allocate(&cs_buf);
-
-        let func = barretenberg
-            .instance
-            .exports
-            .get_function("acir_proofs_get_exact_circuit_size")
-            .unwrap();
-
-        let params: Vec<_> = vec![cs_ptr.clone()];
-        match func.call(&params) {
-            Ok(vals) => {
-                let i32_bytes = vals.first().cloned().unwrap().unwrap_i32().to_be_bytes();
-                let u32_val = u32::from_be_bytes(i32_bytes);
-                barretenberg.free(cs_ptr);
-                u32_val
-            }
-            Err(_) => {
-                unreachable!("failed on acir_proofs_get_exact_circuit_size call");
-            }
-        }
-    }
-
-    pub fn compute_proving_key(&mut self) -> Vec<u8> {
+    pub(crate) fn compute_proving_key(&mut self) -> Vec<u8> {
         let cs_buf = self.constraint_system.to_bytes();
-        let cs_ptr = self.barretenberg.allocate(&cs_buf);
 
-        let pk_size = self
-            .barretenberg
-            .call_multiple(
-                "acir_proofs_init_proving_key",
-                vec![&cs_ptr, &Value::I32(0)],
-            )
-            .value();
+        cfg_if::cfg_if! {
+            if #[cfg(feature = "native")] {
+                let mut pk_addr: *mut u8 = std::ptr::null_mut();
+                let pk_ptr = &mut pk_addr as *mut *mut u8;
 
-        let pk_ptr = self.barretenberg.slice_memory(0, 4);
-        let pk_ptr = u32::from_le_bytes(pk_ptr[0..4].try_into().unwrap());
+                let pk_size;
+                unsafe {
+                    pk_size = barretenberg_sys::composer::init_proving_key(&cs_buf, pk_ptr);
+                }
 
-        self.barretenberg.slice_memory(
-            pk_ptr as usize,
-            pk_ptr as usize + pk_size.unwrap_i32() as usize,
-        )
+                std::mem::forget(cs_buf);
+
+                let result;
+                unsafe {
+                    result = Vec::from_raw_parts(pk_addr, pk_size, pk_size);
+                }
+                result
+            } else {
+                use wasmer::Value;
+
+                let cs_ptr = self.barretenberg.allocate(&cs_buf);
+
+                let pk_size = self
+                    .barretenberg
+                    .call_multiple(
+                        "acir_proofs_init_proving_key",
+                        vec![&cs_ptr, &Value::I32(0)],
+                    )
+                    .value();
+
+                let pk_ptr = self.barretenberg.slice_memory(0, 4);
+                let pk_ptr = u32::from_le_bytes(pk_ptr[0..4].try_into().unwrap());
+
+                self.barretenberg.slice_memory(
+                    pk_ptr as usize,
+                    pk_ptr as usize + pk_size.unwrap_i32() as usize,
+                )
+            }
+        }
     }
 
-    pub fn compute_verification_key(&mut self, proving_key: &[u8]) -> Vec<u8> {
-        let g2_ptr = self.barretenberg.allocate(&self.crs.g2_data);
+    pub(crate) fn compute_verification_key(&mut self, proving_key: &[u8]) -> Vec<u8> {
+        cfg_if::cfg_if! {
+            if #[cfg(feature = "native")] {
+                let mut vk_addr: *mut u8 = std::ptr::null_mut();
+                let vk_ptr = &mut vk_addr as *mut *mut u8;
+                let g2_clone = self.crs.g2_data.clone();
+                let pippenger_ptr = self.pippenger.pointer();
+                let proving_key = proving_key.to_vec();
 
-        let pk_ptr = self.barretenberg.allocate(proving_key);
+                let vk_size;
+                unsafe {
+                    vk_size = barretenberg_sys::composer::init_verification_key(
+                        pippenger_ptr,
+                        &g2_clone,
+                        &proving_key,
+                        vk_ptr,
+                    )
+                }
 
-        let vk_size = self
-            .barretenberg
-            .call_multiple(
-                "acir_proofs_init_verification_key",
-                vec![&self.pippenger.pointer(), &g2_ptr, &pk_ptr, &Value::I32(0)],
-            )
-            .value();
+                std::mem::forget(g2_clone);
+                std::mem::forget(proving_key);
 
-        let vk_ptr = self.barretenberg.slice_memory(0, 4);
-        let vk_ptr = u32::from_le_bytes(vk_ptr[0..4].try_into().unwrap());
+                let result;
+                unsafe {
+                    result = Vec::from_raw_parts(vk_addr, vk_size, vk_size);
+                }
+                result.to_vec()
+            } else {
+                use wasmer::Value;
 
-        self.barretenberg.slice_memory(
-            vk_ptr as usize,
-            vk_ptr as usize + vk_size.unwrap_i32() as usize,
-        )
+                let g2_ptr = self.barretenberg.allocate(&self.crs.g2_data);
+
+                let pk_ptr = self.barretenberg.allocate(proving_key);
+
+                let vk_size = self
+                    .barretenberg
+                    .call_multiple(
+                        "acir_proofs_init_verification_key",
+                        vec![&self.pippenger.pointer(), &g2_ptr, &pk_ptr, &Value::I32(0)],
+                    )
+                    .value();
+
+                let vk_ptr = self.barretenberg.slice_memory(0, 4);
+                let vk_ptr = u32::from_le_bytes(vk_ptr[0..4].try_into().unwrap());
+
+                self.barretenberg.slice_memory(
+                    vk_ptr as usize,
+                    vk_ptr as usize + vk_size.unwrap_i32() as usize,
+                )
+            }
+        }
     }
 
-    pub fn create_proof_with_pk(
+    pub(crate) fn create_proof_with_pk(
         &mut self,
         witness: WitnessAssignments,
         proving_key: &[u8],
     ) -> Vec<u8> {
         let cs_buf = self.constraint_system.to_bytes();
-        let cs_ptr = self.barretenberg.allocate(&cs_buf);
-
         let witness_buf = witness.to_bytes();
-        let witness_ptr = self.barretenberg.allocate(&witness_buf);
 
-        let g2_ptr = self.barretenberg.allocate(&self.crs.g2_data);
+        cfg_if::cfg_if! {
+            if #[cfg(feature = "native")] {
+                let mut proof_addr: *mut u8 = std::ptr::null_mut();
+                let p_proof = &mut proof_addr as *mut *mut u8;
+                let g2_clone = self.crs.g2_data.clone();
+                let proving_key = proving_key.to_vec();
 
-        let pk_ptr = self.barretenberg.allocate(proving_key);
+                let proof_size;
+                unsafe {
+                    proof_size = barretenberg_sys::composer::create_proof_with_pk(
+                        self.pippenger.pointer(),
+                        &g2_clone,
+                        &proving_key,
+                        &cs_buf,
+                        &witness_buf,
+                        p_proof,
+                    );
+                }
 
-        let proof_size = self
-            .barretenberg
-            .call_multiple(
-                "acir_proofs_new_proof",
-                vec![
-                    &self.pippenger.pointer(),
-                    &g2_ptr,
-                    &pk_ptr,
-                    &cs_ptr,
-                    &witness_ptr,
-                    &Value::I32(0),
-                ],
-            )
-            .value();
+                std::mem::forget(g2_clone);
+                std::mem::forget(proving_key);
+                std::mem::forget(cs_buf);
+                std::mem::forget(witness_buf);
 
-        let proof_ptr = self.barretenberg.slice_memory(0, 4);
-        let proof_ptr = u32::from_le_bytes(proof_ptr[0..4].try_into().unwrap());
+                let result;
+                unsafe {
+                    result = Vec::from_raw_parts(proof_addr, proof_size, proof_size);
+                }
+            } else {
+                use wasmer::Value;
 
-        let proof = self.barretenberg.slice_memory(
-            proof_ptr as usize,
-            proof_ptr as usize + proof_size.unwrap_i32() as usize,
-        );
-        proof::remove_public_inputs(self.constraint_system.public_inputs.len(), &proof)
+                let cs_ptr = self.barretenberg.allocate(&cs_buf);
+                let witness_ptr = self.barretenberg.allocate(&witness_buf);
+                let g2_ptr = self.barretenberg.allocate(&self.crs.g2_data);
+                let pk_ptr = self.barretenberg.allocate(proving_key);
+
+                let proof_size = self
+                    .barretenberg
+                    .call_multiple(
+                        "acir_proofs_new_proof",
+                        vec![
+                            &self.pippenger.pointer(),
+                            &g2_ptr,
+                            &pk_ptr,
+                            &cs_ptr,
+                            &witness_ptr,
+                            &Value::I32(0),
+                        ],
+                    )
+                    .value();
+
+                let proof_ptr = self.barretenberg.slice_memory(0, 4);
+                let proof_ptr = u32::from_le_bytes(proof_ptr[0..4].try_into().unwrap());
+
+                let result = self.barretenberg.slice_memory(
+                    proof_ptr as usize,
+                    proof_ptr as usize + proof_size.unwrap_i32() as usize,
+                );
+            }
+        }
+
+        proof::remove_public_inputs(self.constraint_system.public_inputs.len(), &result)
     }
 
-    pub fn verify_with_vk(
+    pub(crate) fn verify_with_vk(
         &mut self,
         // XXX: Important: This assumes that the proof does not have the public inputs pre-pended to it
         // This is not the case, if you take the proof directly from Barretenberg
@@ -196,46 +349,54 @@ impl StandardComposer {
         // from proofs created by Barretenberg. Then in Verify we prepend them again.
         //
         let proof = proof::prepend_public_inputs(proof.to_vec(), public_inputs);
-
         let cs_buf = self.constraint_system.to_bytes();
-        let cs_ptr = self.barretenberg.allocate(&cs_buf);
 
-        let proof_ptr = self.barretenberg.allocate(&proof);
+        cfg_if::cfg_if! {
+            if #[cfg(feature = "native")] {
+                let verification_key = verification_key.to_vec();
 
-        let g2_ptr = self.barretenberg.allocate(&self.crs.g2_data);
+                let verified;
+                unsafe {
+                    verified = barretenberg_sys::composer::verify_with_vk(
+                        &self.crs.g2_data,
+                        &verification_key,
+                        &cs_buf,
+                        &proof,
+                    );
+                }
+                verified
+            } else {
+                use wasmer::Value;
 
-        let vk_ptr = self.barretenberg.allocate(verification_key);
+                let cs_ptr = self.barretenberg.allocate(&cs_buf);
+                let proof_ptr = self.barretenberg.allocate(&proof);
+                let g2_ptr = self.barretenberg.allocate(&self.crs.g2_data);
+                let vk_ptr = self.barretenberg.allocate(verification_key);
 
-        let verified = self
-            .barretenberg
-            .call_multiple(
-                "acir_proofs_verify_proof",
-                vec![
-                    &g2_ptr,
-                    &vk_ptr,
-                    &cs_ptr,
-                    &proof_ptr,
-                    &Value::I32(proof.len() as i32),
-                ],
-            )
-            .value();
+                let verified = self
+                    .barretenberg
+                    .call_multiple(
+                        "acir_proofs_verify_proof",
+                        vec![
+                            &g2_ptr,
+                            &vk_ptr,
+                            &cs_ptr,
+                            &proof_ptr,
+                            &Value::I32(proof.len() as i32),
+                        ],
+                    )
+                    .value();
 
-        self.barretenberg.free(proof_ptr);
+                self.barretenberg.free(proof_ptr);
 
-        match verified.unwrap_i32() {
-            0 => false,
-            1 => true,
-            _ => panic!("Expected a 1 or a zero for the verification result"),
+                match verified.unwrap_i32() {
+                    0 => false,
+                    1 => true,
+                    _ => panic!("Expected a 1 or a zero for the verification result"),
+                }
+            }
         }
     }
-}
-
-fn pow2ceil(v: u32) -> u32 {
-    let mut p = 1;
-    while p < v {
-        p <<= 1;
-    }
-    p
 }
 
 #[cfg(test)]
@@ -247,7 +408,7 @@ mod test {
     #[test]
     fn test_no_constraints_no_pub_inputs() {
         let constraint_system = ConstraintSystem {
-            var_num: 0,
+            var_num: 4,
             public_inputs: vec![],
             logic_constraints: vec![],
             range_constraints: vec![],
@@ -622,6 +783,8 @@ mod test {
 
         test_composer_with_pk_vk(constraint_system, vec![case_1]);
     }
+
+    // TODO: Add logic constaint test (copy from C++)
 
     #[derive(Clone, Debug)]
     struct WitnessResult {
