--- barretenberg_static_lib/Cargo.toml
@@ -1,18 +1,0 @@
-[package]
-name = "barretenberg_static_lib"
-version = "0.1.0"
-authors.workspace = true
-license.workspace = true
-edition.workspace = true
-rust-version.workspace = true
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-
-common = { path = "../common" }
-barretenberg-sys = "0.1.1"
-
-[dev-dependencies]
-blake2 = "0.9.1"
-tempfile = "*"
