--- barretenberg_wasm/README.md
@@ -1,8 +1,0 @@
-
-# barretenberg_wasm
-
-`barretenberg_wasm` is a Noir backend which uses Aztec's Barretenberg proving system through a wasm interface. 
-
-## `barretenberg.wasm`
-
-`barretenberg_wasm` interacts with Barretenberg through the wasm blob `barretenberg.wasm`. This is currently built in [noir-lang/aztec-connect](https://github.com/noir-lang/aztec-connect/tree/kw/noir-dsl/barretenberg) and manually copied across.
