--- barretenberg_wasm/src/acvm_interop/mod.rs
@@ -1,7 +1,0 @@
-pub mod proof_system;
-pub mod pwg;
-
-mod smart_contract;
-pub struct Plonk;
-
-impl common::acvm::Backend for Plonk {}
